+++ CHANGELOG.md
@@ -1,10 +1,17 @@
 # Changelog
 
-# v0.3 (2024-05-08)
+## v0.4 (2024-05-12)
+- ability to use `with` function with spinner
+- added ability to use decorators
+- re-named spinners:
+  - `digit` is now `digits`
+  - `circleDigit` is now `circleDigits`
+
+## v0.3 (2024-05-08)
 - new spinners: newton
 - renamed spinner `spinner` to `slash`
 
-# v0.2 (2024-05-01)
+## v0.2 (2024-05-01)
 - switched to camelCase for spinner names
 - new spinners
 - cursor hiding ability (default)